Move some functionality into a base EventCreator class
