+++ opsdroid/connector/slack/events.py
@@ -44,14 +44,14 @@ class Blocks(events.Message):
             self.blocks = json.dumps(self.blocks)
 
 
-class SlackEventCreator:
+class SlackEventCreator(events.EventCreator):
     """Create opsdroid events from Slack ones."""
 
-    def __init__(self, connector):
+    def __init__(self, connector, *args, **kwargs):
         """Initialise the event creator"""
+        super().__init__(connector, *args, **kwargs)
         self.connector = connector
 
-        self.event_types = defaultdict(lambda: self.skip)
         self.event_types['message'] = self.create_room_message
 
         self.message_events = defaultdict(lambda: self.skip)
@@ -61,19 +61,10 @@ class SlackEventCreator:
             }
         )
 
-    async def create_event(self, event, channel):
-        """Dispatch any Slack event."""
-        return await self.event_types[event["type"]](event, channel)
-
     async def create_room_message(self, event, channel):
         """Dispatch a message event of arbitrary subtype."""
         msgtype = event['subtype'] if 'subtype' in event.keys() else 'message'
         return await self.message_events[msgtype](event, channel)
-
-    @staticmethod
-    async def skip(event, roomid):
-        """Do not handle this event type"""
-        return None
 
     async def create_message(self, event, channel):
         """Send a Message event."""
